packages/backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
python-multipart==0.0.9
supabase==2.50.3
openai-whisper==20231117
SpeechRecognition==3.10.0
pydub==0.25.1
python-dotenv==1.0.0
pydantic==2.11.7 